pyproject.toml
[project]
name = "ccbench"
version = "0.1.0"
description = "A versatile evaluation suite for coding agents."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml",
    "tomli-w>=1.2.0",
]

[tool.ruff]
line-length = 94
